Fix: Display error when attempting to reset nonexistent due date extension
